fix: made sub components
