+++ src/api/lead.ts
@@ -15,7 +15,6 @@ export function useGetLeads(id: string, params: ILeadsParams) {
   const memoizedValue = useMemo(() => {
     const leads = {
       data: data?.data.map((lead: ILead) => ({
-        // ...lead,
         leadId: lead.id,
         status: lead.status,
         source: lead.source,
